refactor: move log_adapter.py under logging/ directory
