+++ bigframes/core/groupby/dataframe_group_by.py
@@ -26,10 +26,10 @@ import pandas as pd
 from bigframes import session
 from bigframes.core import agg_expressions
 from bigframes.core import expression as ex
-from bigframes.core import log_adapter
 import bigframes.core.block_transforms as block_ops
 import bigframes.core.blocks as blocks
 from bigframes.core.groupby import aggs, group_by, series_group_by
+from bigframes.core.logging import log_adapter
 import bigframes.core.ordering as order
 import bigframes.core.utils as utils
 import bigframes.core.validations as validations
