+++ bigframes/ml/ensemble.py
@@ -23,7 +23,7 @@ import bigframes_vendored.sklearn.ensemble._forest
 import bigframes_vendored.xgboost.sklearn
 from google.cloud import bigquery
 
-from bigframes.core import log_adapter
+from bigframes.core.logging import log_adapter
 import bigframes.dataframe
 from bigframes.ml import base, core, globals, utils
 import bigframes.session
