+++ bigframes/operations/structs.py
@@ -17,7 +17,8 @@ from __future__ import annotations
 import bigframes_vendored.pandas.core.arrays.arrow.accessors as vendoracessors
 import pandas as pd
 
-from bigframes.core import backports, log_adapter
+from bigframes.core import backports
+from bigframes.core.logging import log_adapter
 import bigframes.dataframe
 import bigframes.operations
 import bigframes.series
